% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookdown_inrae.R
\name{bookdown_inrae}
\alias{bookdown_inrae}
\title{Template Bookdown respectant la charte graphique INRAE}
\usage{
bookdown_inrae(dir = "MyProject")
}
\arguments{
\item{dir}{Le dossier dans lequel copier le projet bookdown}
}
\value{

}
\description{
Template Bookdown respectant la charte graphique INRAE
}
