% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_analysis_dir.R
\name{create_analysis_dir}
\alias{create_analysis_dir}
\title{Create analysis directory template}
\usage{
create_analysis_dir(dir = getwd())
}
\arguments{
\item{dir}{the directory where the template will be created}
}
\value{

}
\description{
Create analysis directory template
}
\examples{
library(InraeThemes)
# create_analysis_dir(getwd())
}
