% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_documents.R
\name{create_presentation}
\alias{create_presentation}
\title{Create a Quarto presentation template with INRAE theme}
\usage{
create_presentation(dir = "inrae_presentation", change_dir = TRUE)
}
\arguments{
\item{dir}{the directory where the template will be copied to}

\item{change_dir}{should the working directory be changed to dir ?}
}
\value{
a quarto presentation template
}
\description{
Create a Quarto presentation template with INRAE theme
}
