% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_datasets}
\alias{example_datasets}
\title{Datasets fictifs pour l'illustration}
\format{
Quatre dataframes
}
\usage{
example_datasets
}
\description{
Cet objet contient une liste avec 4 datasets servant a l'illustration des themes.
}
\keyword{datasets}
