% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inrae_report.R
\name{inrae_report}
\alias{inrae_report}
\title{Create analysis directory template}
\usage{
inrae_report(dir = "MyProject")
}
\arguments{
\item{dir}{the directory where the analysis template will be created}
}
\value{

}
\description{
Create analysis directory template
}
\examples{
library(InraeThemes)
\dontrun{new_analysis()}
}
