% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_documents.R
\name{latex_presentation}
\alias{latex_presentation}
\title{Copy a LaTeX template for presentations}
\usage{
latex_presentation(dir = "MyProject")
}
\arguments{
\item{dir}{the directory where the template will be copied to}
}
\value{
a full latex presentation template
}
\description{
Template proposed by E. Quinton (INRAE/EABX)
}
