% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_documents.R
\name{latex_report}
\alias{latex_report}
\title{Copy a LaTeX template for reports}
\usage{
latex_report(dir = "MyProject")
}
\arguments{
\item{dir}{the directory where the template will be copied to}
}
\value{
a full latex template
}
\description{
Template proposed by E. Quinton (INRAE/EABX)
}
