% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_analysis_dir.R
\name{new_analysis}
\alias{new_analysis}
\title{Architecture type d'une analyse}
\usage{
new_analysis(dir = "MyProject")
}
\arguments{
\item{dir}{Le nom du dossier}
}
\value{

}
\description{
Architecture type d'une analyse
}
\examples{
library(InraeThemes)
\dontrun{new_analysis()}
}
