% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{palette_inrae}
\alias{palette_inrae}
\title{Palette de couleurs INRAE}
\usage{
palette_inrae()
}
\value{
un vecteur contenant les couleurs
}
\description{
Cette palette correspond aux couleurs issues de la charte INRAE V3.
}
