% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_inrae}
\alias{scale_color_inrae}
\title{Echelle de couleur INRAE}
\usage{
scale_color_inrae(...)
}
\arguments{
\item{...}{Autres arguments pour \code{discrete_scale}.}
}
\value{
a color scale
}
\description{
Echelle de couleur INRAE
}
