% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_inrae}
\alias{scale_fill_inrae}
\title{INRAE fill palette}
\usage{
scale_fill_inrae(...)
}
\arguments{
\item{...}{Other arguments to \code{discrete_scale}.}
}
\value{
a fill scale
}
\description{
INRAE fill palette
}
