% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_inrae_gt}
\alias{theme_inrae_gt}
\title{gt theme with INRAE colors}
\usage{
theme_inrae_gt(gt_table)
}
\arguments{
\item{gt_table}{a gt table}
}
\value{
a gt theme
}
\description{
This gt theme uses colors and fonts defined in INRAE's design system
}
