% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_quant}
\alias{theme_quant}
\title{Title}
\usage{
theme_quant(base_size = 15, base_family = "Avenir Next Pro")
}
\arguments{
\item{base_size}{base size}

\item{base_family}{base family}
}
\value{

}
\description{
Title
}
