% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_documents.R
\name{create_book}
\alias{create_book}
\title{Create a Quarto report/book template with INRAE theme}
\usage{
create_book(dir = "quarto-inrae-book", change_dir = TRUE)
}
\arguments{
\item{dir}{the directory where the template will be copied to}

\item{change_dir}{should the working directory be changed to dir ?}
}
\value{
a quarto book template
}
\description{
Create a Quarto report/book template with INRAE theme
}
