% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_analysis_project.R
\name{new_analysis}
\alias{new_analysis}
\title{Architecture type d'une analyse
Inpired by \code{ProjectTemplate}}
\usage{
new_analysis(projname = "MyProject", init = TRUE)
}
\arguments{
\item{projname}{Name of the project to be created}

\item{init}{Should a git repository be initialised ?}
}
\value{
a data analysis directory
}
\description{
Architecture type d'une analyse
Inpired by \code{ProjectTemplate}
}
