% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_analysis_dir.R
\name{new_analysis}
\alias{new_analysis}
\title{Create analysis directory template}
\usage{
new_analysis(dir = "MyProject")
}
\arguments{
\item{dir}{the directory where the analysis template will be created}
}
\value{

}
\description{
Create analysis directory template
}
\examples{
library(InraeThemes)
\dontrun{new_analysis()}
}
