% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{palette_inrae}
\alias{palette_inrae}
\title{INRAE color palette}
\usage{
palette_inrae()
}
\value{
a named vector of colors
}
\description{
This color palette is based on colors from INRAE's design system.
}
