% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_inrae}
\alias{scale_color_inrae}
\title{INRAE color palette}
\usage{
scale_color_inrae(...)
}
\arguments{
\item{...}{Other arguments to \code{discrete_scale}.}
}
\value{
a color scale
}
\description{
INRAE color palette
}
