% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_inrae}
\alias{scale_fill_inrae}
\title{Echelle de remplissage INRAE}
\usage{
scale_fill_inrae(...)
}
\arguments{
\item{...}{Autres arguments pour \code{discrete_scale}.}
}
\value{
a fill scale
}
\description{
Echelle de remplissage INRAE
}
