% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_inrae}
\alias{theme_inrae}
\title{Theme ggplot aux couleurs INRAE}
\usage{
theme_inrae(base_size = 15)
}
\arguments{
\item{base_size}{base size}
}
\value{
a ggplot theme
}
\description{
Ce theme ggplot correspond aux couleurs issues de la charte graphique INRAE V3.
}
