% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_inrae}
\alias{theme_inrae}
\title{ggplot theme with INRAE colors}
\usage{
theme_inrae(base_size = 12)
}
\arguments{
\item{base_size}{base size}
}
\value{
a ggplot theme
}
\description{
This ggplot theme uses colors and fonts defined in INRAE's design system
}
